Update README.md
Forsage is a very good networking place. Forsage is getting a bad name due to some unscrupulous leaders. They are giving wrong information to people to buy slots and common people are suffering. So forsage means cheating is spreading all around. But does Forsage cheat? Of course not. Will the force not take any action against those cheaters? As we've come to expect from Forsage, cheaters are punished. And want to forsage with respect